move common controls to ParticleStyle and leave mesh specific stuff in MeshSetupStyle
